+++ src/Repository.ts
@@ -25,15 +25,20 @@ class Repository<T> implements IRepository<T> {
     this._store = store;
   }
 
-  public async save(id: string, events: IDomainEvent[], version: number) {
+  public async save(
+    id: string,
+    events: IDomainEvent[],
+    version: number,
+    metadata?: object
+  ) {
     const aggregateId = this._getAggregateId(id);
-    return this._store.save(aggregateId, events, version);
+    return this._store.save(aggregateId, events, version, metadata);
   }
 
   public async getById(id: string): Promise<IAggregateState<T>> {
     const aggregateId = this._getAggregateId(id);
     const events = await this._store.loadEvents(aggregateId);
-    return this._createInstance(events);
+    return this._createInstance(id, events);
   }
 
   public async getNextId(): Promise<string> {
@@ -44,9 +49,12 @@ class Repository<T> implements IRepository<T> {
     return { id, name: this._aggregate.name };
   }
 
-  private _createInstance(events: IAggregateEvent[]): IAggregateState<T> {
-    const { state, version } = this._aggregate.rehydrate(events);
-    return { state, version, exists: version > 0 };
+  private _createInstance(
+    id: string,
+    events: IAggregateEvent[]
+  ): IAggregateState<T> {
+    const { state, version } = this._aggregate.rehydrate(id, events);
+    return { id, state, version, exists: version > 0 };
   }
 }