+++ src/__tests__/AggregateRoot.spec.ts
@@ -83,7 +83,7 @@ test('simple command handler', async t => {
   const incrementedByOne = createCommand('increment', 0);
 
   const events = await counterAggregate.applyCommand(
-    counterAggregate.initialState,
+    counterAggregate.getInitialState('test'),
     incrementedByOne
   );
 
@@ -93,7 +93,10 @@ test('simple command handler', async t => {
 test('handle domain error', async t => {
   const command = createCommand('incrementByPositive', 0, { step: -2 });
   const shouldThrow = () =>
-    counterAggregate.applyCommand(counterAggregate.initialState, command);
+    counterAggregate.applyCommand(
+      counterAggregate.getInitialState('test'),
+      command
+    );
 
   await t.throwsAsync(shouldThrow, { name: 'DomainError' });
 });
@@ -104,7 +107,7 @@ test('multiple yielding command handler', async t => {
   });
 
   const events = await counterAggregate.applyCommand(
-    counterAggregate.initialState,
+    counterAggregate.getInitialState('test'),
     incrementedByDynamic
   );
 
@@ -120,7 +123,10 @@ test('multiple yielding command handler', async t => {
 test('command validation', async t => {
   const badCommand = createCommand('incrementByDynamic', 0, { steps: 'foo' });
   const shouldThrow = () =>
-    counterAggregate.applyCommand(counterAggregate.initialState, badCommand);
+    counterAggregate.applyCommand(
+      counterAggregate.getInitialState('test'),
+      badCommand
+    );
 
   await t.throwsAsync(shouldThrow, {
     instanceOf: CommandValidationError
@@ -130,6 +136,7 @@ test('command validation', async t => {
 test('stateful multiple yielding', async t => {
   const initial = {
     exists: true,
+    id: 'test',
     state: { value: 5 },
     version: 1
   };
@@ -148,7 +155,7 @@ test('stateful multiple yielding', async t => {
 });
 
 test('asynchronous commands', async t => {
-  const initial = counterAggregate.initialState;
+  const initial = counterAggregate.getInitialState('test');
   const delayedIncrement = createCommand('delayedIncrement', 0);
   const events = await counterAggregate.applyCommand(initial, delayedIncrement);
 
@@ -157,7 +164,7 @@ test('asynchronous commands', async t => {
 });
 
 test('asynchronous error handling', async t => {
-  const initial = counterAggregate.initialState;
+  const initial = counterAggregate.getInitialState('test');
   const delayedIncrement = createCommand('delayedIncrement', 0, { step: -1 });
 
   const shouldThrow = counterAggregate.applyCommand(initial, delayedIncrement);
@@ -166,7 +173,7 @@ test('asynchronous error handling', async t => {
 });
 
 test('asynchronous yielding command handler', async t => {
-  const initial = counterAggregate.initialState;
+  const initial = counterAggregate.getInitialState('test');
   const delayedIncrement = createCommand('delayedYieldingIncrement', 0);
   const events = await counterAggregate.applyCommand(initial, delayedIncrement);
 
@@ -175,7 +182,7 @@ test('asynchronous yielding command handler', async t => {
 });
 
 test('multiple command handlers', async t => {
-  const initial = counterAggregate.initialState;
+  const initial = counterAggregate.getInitialState('test');
   const arrayIncrement = createCommand('arrayIncrements', 0);
   const events = await counterAggregate.applyCommand(initial, arrayIncrement);
   t.is(events.length, 3);